IDEA-330376 Use invokeLater instead of invoke in BgtTreeWalker
Turns out sometimes onSuccess is invoked immediately on
the calling thread. It's not clear why that happens, perhaps
the background thread is executing requests very quickly.
When that happens, it may lead (and sometimes leads) to a stack overflow.

Fix by using invokeLater to clear the stack and ensure that
visiting is executed in a sequential manner and not in a reentrant one.

This may or may not fix the issue, but at least it should take care
of stack overflows and make further debugging easier.

GitOrigin-RevId: 275764d65050a66483ff72777bee74d7e125edd0