IDEA-327405 UI Inspector: save stacktraces only if it is explicitly enabled
`Added-at` stacktraces are needed much rarely, so there is no need to save them right after the first invocation of UI Inspector. The users that need the stacktraces often will just enable it by default. Others will need to enable it in the inspector dialog for the current IDE session.

GitOrigin-RevId: fac89446caafaf819e568943f690dacfdceaa03a