EA-917083 Catch exceptions in ImageFileService
Exception handling code was accidentally lost when
converting to coroutines. Catch and log everything
to avoid uncaught exceptions. Ignore cancellation
exceptions because they're expected and not
errors. We don't have to reset the image on
cancellation either, as it could cause unnecessary
flickering when it's cancelled due to another
request arriving (another file is being loaded).

GitOrigin-RevId: 7ed4058783bd1d650a18b26eb179586021c9f604