IDEA-333849 When Select Opened File is clicked, check for detached editors
This is yet another case when the technical implementation makes sense,
but doesn't correspond to the user's expectations.

When the Select Opened File action button is clicked, the IDE frame
is obviously focused, and so is the selected editor in the same frame.
This makes the action to use that editor to find the file to select
in the project view. However, if there's a detached editor opened
in another window, and that window was active immediately before
the user clicked the action button, the user obviously expects
the file opened in THAT window to be selected.

To fix, we keep track of the last two focused editors, and also
keep track of how many mouse clicks (actually, releases)
the IDE frame has received since it was activated the last time.
If the action is activated by clicking the action button, and if that
is the first click since window activation, we check both
the latest focused and the previously focused editors.
It's needed because we don't know whether the editor
has processed the window activation event yet
(in my testing it doesn't). Then we pick the editor that is
not from the current window, but was just active.

We use weak references to keep track of the editors
and make sure to clear them if the editor is disposed,
to avoid dereferencing them again and again
(and thus creating strong references every time).

All of this crazy stuff only happens when it's the first
click since window activation, so at least it shouldn't affect
normal IDE use in any way.

GitOrigin-RevId: 23f5ad75573926ed9fab3539e5cc8ea560799e88