[platform] skip loading unresolved content module if it's marked as optional in product-modules.xml (GTW-6350)
'JetBrainsClient.xml' includes 'intellij.cwm.guest.notebooks' as a content module, but it depends on 'intellij.notebooks.ui' module which isn't present in all IDEs. So in order to start JetBrains Client from installations of such IDEs, we need to skip this module. It's possible to achieve this by adding a new plugin which includes 'intellij.cwm.guest.notebooks' as an optional part, but adding a new module and a new plugin looks too much for that. Adding 'intellij.cwm.guest.notebooks' to product-modules.xml as an optional does the trick, but until ModuleBasedProductLoadingStrategy can load descriptors for modules from product-modules.xml directly, we need to skip this module if it's absent in ClassPathXmlPathResolver.

GitOrigin-RevId: 69b8c7443cdfd8d537fabe37d7dbd44164b46751