[gitlab] Snippets: Fix errors thrown from `getCurrentUser` check
Exceptions thrown from this check are indicative of some problem
with the login details or with the connection to a server.
Without catching them, they end up being reported, which is unwanted
in this case. We just want to re-attempt login if some problem
occurs.

In this case, I think we ignore the possibility of connection errors
as there's no easy way to recover from those.

GitOrigin-RevId: 85515ef1973693b7153c20baafd449415f14bdea