IDEA-333971 AndroidGradleProjectStartupActivity should wait for ExternalProjectsManager to finish initialization
ExternalSystemStartupActivity and AndroidGradleProjectStartupActivity used to run in an order previously.
But recently IntelliJ platform started running post-startup activities concurrently.

Because of that, the old contract "ExternalProjectManager is ready when AndroidGradleProjectStartupActivity starts"
is not true anymore. Thus, AndroidGradleProjectStartupActivity should explicitly wait for ExternalProjectManager to
be initialized.

ExternalProjectManager has `runWhenInitialized` method for that. But, unfortunately, it causes a deadlock in tests
because it runs `runnable` on EDT which is waiting for the project opening to finish.
To overcome this issue, I've added `runWhenInitializedInBackground` method.

GitOrigin-RevId: c616ad179652686a9827de4496d67e6e152d8f52