IDEA-282484 Add another tool window border if the existing one is equal to background
The implementation of this consists of two parts.

First, components with editor-like backgrounds are stored in a list
inside InternalDecoratorImpl. So far I've found three of these:
editor text fields (Commit), consoles (Run, Debug, etc.) and terminal.
Whenever such a component is added/removed (addNotify/removeNotify),
it looks up an InternalDecoratorImpl among its parents, and if one
exists, it registers / de-registers there. This step is needed to avoid
traversing the whole hierarchy on every paint operation.

Then, when it's time to compute border insets and/or paint the border,
a whole bunch of cases are considered. The bottom line is,
whenever a window has a component with editor-like background,
and this component touches the edge that borders the editor
or another tool window with a similar component along the common edge,
then the additional border is painted. In case of two tool windows touching,
the border is painted only on one side (doesn't really matter which),
as a sandwich of borders and the splitter divider looks ugly otherwise.

This whole logic only kicks in when the usual border color equals to
the default editor background. This prevents it from kicking in
for the light theme or for custom themes that specify custom border
colors different from the editor background.

This can probably be improved by considering the actual colors
being used, as the editor scheme can be set up to have the console
background color to be different from the default editor background,
but that would complicate the implementation even further
and is not really necessary with the default themes and schemes.

GitOrigin-RevId: b83fc8dee65071612537d3801e094cf51533d19a