IDEA-332466 Focus the Got It tooltip for Speed Search
This turned out trickier than it would seem. The reason why pressing Enter
would close both the Got It tool tip and the Speed Search field was because
the Got It tool tip never received focus to begin with, so pressing Enter
would be processed directly by the tree and Speed Search, which caused
the Speed Search to disappear first, and the Got It tool tip to disappear
then as a consequence.

So, make the Got It tool tip focused, problem solved? Well, no.
In that case, the Speed Search loses focus and that causes it to
close immediately, and that also hides the Got It tool tip as well.
The result is that neither of them effectively shows up.

To fix this for good, the Speed Search had to be modified
so that it no longer closes on losing focus if there's a Got It tool tip
shown for it, and the Got It tool tip API had to be modified to,
first of all, add the method to check if a Got It tool tip is shown
for a component, and then the implementation had to be modified
as well to set the BALLOON_PROPERTY earlier, right when the balloon
is created, not after it's shown, because otherwise it isn't set yet
when focus is requested, and the Speed Search popup still believes
that no Got It tool tip is shown.

GitOrigin-RevId: 3a64b27e880589629b97cd80641e446144d4013d