(IDEA-332279) ML in SE: use single RA to find search results and process them
The problem with deadlock is still resolved since we wait on condition outside the RA. However, the returned search candidates can be outdated before processing, so it's better to filter and process them under a single read lock. I also add read access assertion in the place where we need it.

GitOrigin-RevId: 64f6c4ddcd35f4db1819d88f00788828591da170