IDEA-332523 Disable Speed Search for the Threads combo box in the debugger
It looks like the combo box had two different types of speed search installed:
the one I removed here and another one that is really invoked when something
is typed. There was a race condition between them, causing the latter to be
used in normal conditions, but the former could still be invoked by setting
suspending break points in SpeedSearchBase on key typed events,
which looked ugly, and probably not what's intended at all.

Looks like the regular speed search is obsolete here and replaced by
the one that is normally invoked, but nobody bothered to remove
the old one. Now it confuses the Speed Search action in the tool window
menu, because it detects the installed speed search (it doesn't care
by the other one, as it uses a different implementation) and tries
to invoke it, when it should actually invoke speed search for the
variables instead.

Fix by simply removing the combo box speed search that isn't
really used anyway.

GitOrigin-RevId: 3b9786ceff3880aa6d486466138ba5cad1b08142