Remove extraneous completion logging
This logging over-fires when coroutines are cancelled, which is standard
practice for the InlineCompletionHandler when new keystrokes/completion
events are received. Similar changes have already been made upstream,
but can't be cherry-picked due to the many conflicting changes in this
area of code.

This change can be removed prior to the 2023.3 merge, and that's tracked
with bug 304298351.

Bug: 304298351
Test: n/a
Change-Id: Ide11d4b603a1d8f4c2fbe0fb4c68b1fd90a702bb