Add basic telemetry to InlineCompletionHandler
This change mimics parts of commit 304299822 from upstream. It cannot be
cherry-picked due to code churn in the area. This change can be reverted
in advance of the 2023.3 merge, and replacement implementations will be
brought in with the merge.

Some of the interfaces and classes will be used in Studio code. The
implementations being added here match what exists now in
intellij-community, so there should not be any build issues at the time
of the merge. (Unless there are further changes of course, but if that
happens we can adjust to it.)

Bug: 304299822
Test: n/a
Change-Id: If2c535932ac00bdb692b896ac34cce39b00f9b69