IDEA-333972 Close BookmarkTypeChooser first thing onChosen
Some code in onChosen may invoke modal dialogs,
to choose the list to add the new bookmark to, for example.
If the popup isn't closed yet at this point, then it may overlap
the newly opened dialog.

Fix by first closing the popup in onChosen, then doing the rest.
It's not like keeping it open makes much sense at that point.

GitOrigin-RevId: 4386377163a7829aaf24eb7db1b84b134b1c80aa