IDEA-333209 Update File Name Widget sooner
By default, the main toolbar is updated with some delay,
using a timer. This doesn't work well for the file name
widget, as we want it to change immediately.

Unfortunately, as there seems to be no easy way to
update just that widget, the whole toolbar is updated for now.

A new service was created to implement these updates.
It's initialized on the first file name widget update and then
just keeps track of the selected / opened editors and
invokes the update immediately.

GitOrigin-RevId: 7c3147f3483d0850fca1ec6b1196d220d73ad607