IDEA-333849 Fix the last focused editor hack for detached project view
First of all, get rid of that ugly focused editor tracking and rely
on FileEditorManagerImpl instead. A new method has been added
to get the last focused editor even if the corresponding registry
key is disabled (which it is by default).

Then, to determine whether the last focused editor should be used
or the old behavior, the following logic is implemented:

1. First of all, if the action wasn't invoked by clicking the tool
window button, the old logic works.

2. If the button was clicked in a detached project view,
then the last focused editor always wins. This is because
it doesn't make any sense to prefer one window with editors
over the other when both are not focused (since the project view
is focused instead).

3. If the project view is attached to the IDE frame, then
the last focused editor is used only in case the IDE frame was
activated by the same mouse click that invoked the action.
This corresponds to the previous version of this hack.

GitOrigin-RevId: 89a328d98ba6eee59a077ff10e9740526ad01b49