[modular loader] introduce ProductMode class to allow loading some plugin modules in specific modes only (IJPL-270)
This change adds ProductMode class which specifies the mode current process is running in. The mode is passed via 'intellij.platform.product.mode' system property, and used in ModuleBasedProductLoadingStrategy to determine which modules should be loaded. Currently, it uses special marker modules in dependencies for that, see ProductModes for details.

Build scripts are modified to use 'FRONTEND' mode for JetBrains Client, and use 'LOCAL_IDE' mode by default for other IDEs.

A new marker module 'intellij.platform.localIde' is added to mark modules which should be loaded in local IDEs only, but should be excluded from an IDE running in frontend mode (JetBrains Client). This module is added as a dependency to 'intellij.terminal.sh' to ensure that it won't be loaded in JetBrains Client (this is needed for GTW-6350).

GitOrigin-RevId: adbef3b8c51e3473cb7564b6e617f50561ca0cff