[modular loader] temporary skip a plugin if its content includes modules from the main module group (GTW-6350)
If an IDE (e.g. IDEA Community) includes a custom variant of 'intellij.performanceTesting.async' where 'intellij.profiler*' modules are bundled with the plugin, these modules will be loaded as part of the main module group when the embedded JetBrains Client starts (because they are referenced from JetBrainsClientPlugin.xml), so an attempt to load them as part of the plugin will fail with assertion in ClassLoaderConfigurator::checkPackagePrefixUniqueness. Let's skip the loading of such plugins for now.

GitOrigin-RevId: e0df43a48431af49ffcc38db1f4ede3ed1394342