[terminal] Fix showing alternate buffer applications in a separate component
`TerminalModel#useAlternateBuffer` was changed after `TerminalTextBuffer#useAlternateBuffer` call. So as a result, the listeners were not called because we were unable to detect the change.
Store our own alternate buffer state in the `TerminalModel` to fix this issue. It always should be the same as the state in the `TerminalTextBuffer`.

GitOrigin-RevId: 74066e356b5ba0e51d2c82968541513d50e138cb