IDEA-333209 Update File Name Widget faster
Instead of relying on action system update (which happens in a timer),
we store the list of currently showing file name widgets (usually it's
just one or zero, but let's cover the general case anyway). We use
a weak list for that to avoid any leaks just in case. Then we just
update them explicitly whenever the currently opened file
changes.

To avoid subclassing JBPanel for no reason, instead of overriding
addNotify/removeNotify, we just subscribe to the "ancestor" property
change which is fired from those methods in JComponent.

GitOrigin-RevId: 68e8827f1f206c9001361ecdd7bb78914c2a1eb4