IDEA-334288: Fix "Descriptor wasn't found for declaration CLASS (KotlinFullClassNameIndex)"
On IDE shutdown unsaved document (ProfileServiceTest.kt) produces TRANSIENT_STATE_CHANGE. This change cannot be processed immediately, therefore ProfileServiceTest.kt goes to so called dirty-files-queue. On the following start indexes for this file are deleted, but IndexingFlag is not cleared. IDE thinks that file is fully indexed, but in fact indexes for this file were deleted.

GitOrigin-RevId: bcf9248205c154d7b7b2bfcee886e29c6853ef86