IDEA-334573 Fix ImageOrColorPreviewService isValid slow op
Turns out even isValid() is a slow op. Thankfully, it wasn't really
necessary there, more like a guard against possible out of sync
events, but even if the editor is attached, the service doesn't
actually use the PSI file right away, and when it does,
it still checks for its validity, so this fix should be harmless.

GitOrigin-RevId: 5ac0b388d6bb3fb594262839c710a0c8bcf0a027