[gradle][IDEA-334192] fix GradleRunAnythingProviderTestCase
`GradleExecuteTaskAction#runGradle` used to run Gradle tasks with `ProgressExecutionMode.IN_BACKGROUND_ASYNC`. In the unit test environment `IN_BACKGROUND_ASYNC` is executed synchronously. Switching to NO_PROGRESS_ASYNC` results in asynchronous execution in the `GradleRunAnythingProviderTestCase#executeAndWait`, and since the semaphore was in the UP state, `Ref<BuildView>` was empty at the time of data access.

GitOrigin-RevId: 7dce658f7629e5fc16ae644a12b27296f8969059