IDEA-331273 Implement logging of tool window layout changes
It wasn't easy to figure out how much logging is enough,
but at least this much should do for most dimensional issues.

Most logging is deliberately done through the ToolWindowManagerImpl's
logger, so it can be enabled with a single category. Separate loggers are
used for tool window pane changes (for docked windows) and for external
decorators (for floating and windowed ones).

The TRACE-level reshape() logging is likely the most important here,
as it allows to track every bounds change through the API,
and in case a move/resize event originates from the outside world,
it's easily distinguished now by the presence of a move/resize
event in the log without the corresponding reshape() call.

GitOrigin-RevId: c3fa4d00e94d3d15eee06c6853fd0f79d6e09cbf