IDEA-230738 Prevent the OS from resizing detached tool windows for no reason
Sometimes the OS resizes a window right after it's shown, or right before.
It's not clear why this happens, but it seems to happen mostly on Linux
and macOS. To prevent this, we now keep track of the bounds set
through the API, which we conveniently already have (ToolWindowExternalDecorator.getBounds),
and after the window is shown, we watch for move/resize events for some time
(currently 100 ms, but adjustable in the Registry) and revert the bounds back
to what we set them to.

This is, of course, nothing more than an ugly workaround, but it'll likely be more reliable
than endless attempts to hunt down the origin of every move/resize in every
possible environment.

GitOrigin-RevId: c0d4ea284d4147efe99b52aeeb8c7e0f30e42cb2