IDEA-334673 Use frame bounds instead of root pane bounds for windowed tool windows
The code for calculating root pane bounds was introduced in IDEA-148849,
when the code for properly handling resize events was introduced.
It wasn't really necessary, however, as saving the bounds works just as well for
the entire frame as well as for the root pane, but in reality using the entire
frame is more reliable. It was always done that way for floating tool windows,
for example, so the new implementation is more consistent here.
And less error-prone because it's very easy to forget to call getRootBounds
somewhere.

GitOrigin-RevId: 2734d6205d10454dbb4bbe8aa70e36654cb56533