[startup] never show dialog to import configs in splitMode (RDCT-593)
If an IDE starts in splitMode (or remoteDevHost mode), an attempt to show dialog before Application is initialized will fail, because com.intellij.platform.impl.toolkit.IdeToolkit.clientInstance will be called before application is initialized. So for now we disable the dialog via system property in splitMode (like it's disabled in remoteDevHost mode in remote-dev/.../launcher.sh).

Other approaches to fix the problem didn't work. If we postpone setting our IdeToolkit as the main toolkit, it causes ClassCastExceptions from events which were fired before the toolkit is set. If we store instance of the original toolkit, and make IdeToolkit to delegate to it if Application isn't initialized yet, it cause ClassCastExceptions in other places where Toolkit.getDefaultToolkit() is casted to specific implementation.

GitOrigin-RevId: bbdbfdad12c2cc0d27dff476269e628b98a9bcfd