IDEA-334819 External project data is not restored in some cases
The issue is caused by createDataClassResolver from serialization.kt. It can find java classes for
deserialization only in plugins that declare their own ExternalSystemManager implementation. This
is not the case for Android plugin (as it uses the platform's GradleManager), so all data nodes
containing android data structures are never properly deserialized.

Also note that the issue is reproduced when running from sources only in dev-build mode.

GitOrigin-RevId: 38c92df55f04b9b756681da064f2aeac4fab3db6