IDEA-333209, IJ-CR-116331 Split FilenameToolbarWidgetProjectUpdateService in two
Use the app-level service for most tasks, but keep the file flow
in the project-level service to avoid leaks and to make the flow
manageable (as the concept of the currently opened file
is per-project by its nature, we don't want updates overwriting
each other for different projects).

This allows us to remove that ugly listener that relied on
the components being added to the visible hierarchy,
which wasn't the case, for example, when the component is
initially invisible, and we want to make it visible when
we open a file for the first time.

GitOrigin-RevId: 3016159b4bc64ee2941bc6b9f5424021e11b70ef