IDEA-334592 Clear MarkupModelImpl cache when a problem disappears
The cause of the bug is that when a problem disappears,
RangeMarkerTree.RMNode.invalidate calls tree.beforeRemove,
which eventually calls MarkupModelImpl.fireBeforeRemoved.
This call propagates to the tree model, which is correctly updated.
However, the existing problems remain cached in MarkupModelImpl,
so when editor tabs are switched back and forth, these cached
problems pop up again.

This is more like a workaround than a proper fix, but it'll have to do
for now.

GitOrigin-RevId: 4f4657adec59cde2b75d208337aab2084d77727d