[terminal] IDEA-334961 Refactor: use our own escape action instead of editor action handlers for EditorEscape action
There were already 3 instances of EditorActionHandler for EditorEscape action. And additional handlers are expected. It is too much to create so many handlers because they affect all editors. And also, code in EditorActionHandlers is complex, requires a lot of code duplication, and it is really easy to introduce a bug that will affect all editors.
So, it is much shorter and safer to create separate action for the Terminal.

GitOrigin-RevId: 103f8638ae40c6eca998203571cdd4c5833541ea