[collab] rework data-model and model-vm mappers
Model can be ID'ed by a data id and updated, but there should be a direct model-viewmodel mapping due to the fact that a new model instance can appear having the same ID as the existing one.
In this case a VM will not be updated to use a new model which is a mistake.

GitOrigin-RevId: 67655ba0beac8655c45ff2dca34432d0db163048