IDEA-334799 Increment activity tracker on tree selection changes
The reason the wrong file was opened with Open Files with Single Click /
Preview Tab is that the corresponding action uses an async data context,
which used outdated cached information because a change in tree
selection isn't perceived as user activity there. So while the new
file is already selected, the action system thinks the old one is
still selected.

Fix by incrementing the activity tracker on every tree selection
change. Not sure if it's the right fix at all, so it'll have to be
reviewed, but it seems safe enough, as tree selections don't
change so often to make it a performance issue.

GitOrigin-RevId: f93e76ba0cfb4c1d3bb4846d686e49219877a4cf