[terminal] IDEA-334961 Fix not working Escape action when command is running
It required improving the key events handling in the TerminalEventDispatcher.
Before, all key events that correspond to the actions from `ACTIONS_TO_SKIP`, were fully handled by the platform even when there was no enabled action for the key event.
Now, the platform tries to execute the action, but if there is no enabled action, we take the key event back and send it to the Terminal process.
So, when command is running, the Escape action now will first deselect the block, close the search if any, and only then send to the Terminal process.

GitOrigin-RevId: 692cef08f29b74f6d10cb5c2d51704de29b52b65