[Workspace Model] [IDEA-330499] Do not fire JDK update after modification of misc.xml if there is no actual change
In some cases, the async file listeners are scheduled under non-blocking read action. If we execute the write action during such read action, it will be interrupted and restarted. And this may go into the loop.
This change may break the loop as we won't throw the jdk even change in case of no changes.

GitOrigin-RevId: 8d143b0f6fe78542306a494361ea07ec8d819f25