EA-918457 Move mutable TodoItemNode fields to presentation
To avoid concurrent access and modification, it's always best
to keep everything in the presentation, which is always
created, updated and only then atomically published.

It would be possible to get away with mutable fields,
but it would be a pain to synchronize them correctly,
and it would be error-prone: for example, an update
can happen between getAdditionalLines()
and getHighlightedRegions() calls, resulting in potential
mismatches or more AOOBEs. It's best to keep
everything consistent and update in one go.

The presentation having mutable structures is fine
because it's not yet published when it's being updated,
so there's no concurrency at that moment.

There's no real need to have these lists concurrent
(copy-on-write) either, but let's keep them that way
for consistency, as PresentationData's lists are
implemented the same way.

GitOrigin-RevId: 4e419472c6292e79e838ea55788bfff7def645d1