IDEA-333209, IJ-CR-116331 Fix unknown.svg appearing in file name widget
The reason for it appearing was that the action system really
wants to make action visible by default, resetting their visibility
even if set to false in the template presentation. This is OK
as long as the presentation gets updated properly, but it takes
some time, so during initialization we see this ugly button icon
in the middle of the toolbar, which just looks ugly.

Thankfully, in this case the fix is very simple: just make
the whole thing invisible if the text is empty, regardless
of what the presentation says.

GitOrigin-RevId: 23e120557cb806da2364b84310bec976b82db602