[Workspace Model] [IDEA-330857] Simplify internal ChangeEntry
ChangeEntitySource and ReplaceAndChangeSource events are removed. They were used when entity source change was a separate operation. Now it's not needed.

Also, there was a test that modified data and entity source in different builders and the merged them using addDiff. This test was added after IDEA-251817.
This is the case when we can keep simultaneous changes for an entity source, however, it was not intended to have this functionality. Thus, it's removed now.

GitOrigin-RevId: 44d15eafeeab1949efa98cccec1ea9f92d9006d9