[gitlab] Make `canAddNotes` a flow
This is because no comments should be added when a draft note is
already part of the discussion (this is the GitLab behaviour in such a case).

It can be discussed whether we should allow adding notes in case a draft note
was already added, but for now it seems most sensible to disable.

When you try to add a draft note to a discussion that already includes one,
the request will return an error, so this is a way to prevent this completely.

GitOrigin-RevId: c8850c69f3e347837dcb5beaf4fc96ca3c498bfe