IDEA-322036 Fix a potential freeze when selecting Run/Debug tabs
Avoid using `JBTabsImpl#findInfo` because if multiple `TabInfo` hold the same tab content component - it is undefined which `TabInfo` to return. And potentially it can cause the cycle like in the stacktrace from the mentioned ticket.
I failed to find any real examples of such behavior, but I did't see any other reasons for this freeze.

GitOrigin-RevId: 1682716f9568ddd8ab24d196bbb9b9ca0ce268e4