[file-history] ensure rename detection is run for all merge commit parents
"git show" command executed for the merge commit only returns the first rename that was found. This can present a problem, for example when the first rename is detected incorrectly and the correct rename to the other parent is thus not found.

This commit adds rename detection for all parents of a merge commit. This way, both correct and incorrect history branches would be shown in the file history.

GitOrigin-RevId: 5c71931cd0bcc4d57618a9afb4b32807795b6dec