IDEA-335159 Update speed search in structure popup
The root cause of the bug is that the tree is both asynchronous
and filtered. So when an invalid string is entered into the search
field, the tree still shows the previous state, which causes
the speed search to incorrectly behave as if there are still
some elements matching. Similarly, when the string is backspaced
to a valid string, when the update is performed, the tree is
still empty, not reflecting the updated filter value.

Fix by doing two things.

First, call refreshSelection() on the speed search not only
when the first matching element is correctly selected,
but also when selection fails (the promise is rejected).
By that time the tree is already updated.

Second, whenever refreshSelection() is called,
update the popup's state if the popup is shown
by invoking its updateSelection() method.
This way the popup correctly changes its color
whenever the selection is refreshed.

GitOrigin-RevId: 3b176dfd238cc0eebf54927b5f17b8cb7ca979e9