[Kotlin] K2: Decompile built-ins from plugin classloader without filtering
Split stub versions for .kotlin_builtins with K1 and K2 IDE to force
reindexing between switches.

The built-ins from plugin classloader are used by default with the lowest
priority in modules of any platform, including common MPP modules.
kotlin-stdlib-common doesn't contain some of the built-in declarations
in .kotlin_metadata files and doesn't have .kotlin_builtins of its own.
Not providing the declarations in the decompiled .kotlin_builtins leads
to unresolved references for some of the built-in classes, like kotlin.Unit.

KT-61757

GitOrigin-RevId: dae0e6fa191987ca54b61a1bbcf0962405115865