[rdct] use shared config directory to store certificates in JetBrains Client (RDCT-629)
Otherwise, the "Untrusted Server's Certificate" dialog may be shown each time the client starts. It may be better to use per-process directories and synchronize changes in them to be more tolerant to concurrent updates, but these paths are passed to external code, so it's rather hard to implement such synchronization. Let's implement a simple solution for now.

GitOrigin-RevId: 93876961856a2353f798592b5f216a992490159c