[rdct] support "Edit Custom VM Options" action for embedded client (RDCT-571)
Now customizable VM options are saved in a separate file in the config directory of the main IDE, and can be edited via "Edit Custom VM Options" action.

When launchers for JetBrains Client will be included in IDE's installation (RDCT-521) this won't be needed, but it will still make sense for IDEs started from sources.

GitOrigin-RevId: aed614bd2dcb098f2a53cd995b8a63bf1633b7e5