IDEA-327970 IntentionAction#isAvailable is called from EDT
More correct fix.
As soon as the EditorNotification is created (in EDT) we ask corresponding IntentionActionProviders for intention actions (also in EDT, they are cheap) and store them to the FileEditor user data. Then, during ShowIntentionAction collectInformation() (in BGT), we call collectActions() and retrieve the intention actions from FileEditor and update them (also in BGT)

GitOrigin-RevId: ad29a77ba3ff86e90f9884b22aadcff3cca37627