[gitlab] Drafts: check server version to add 'save as draft' option
For this, we remove the `asDraft` parameter, instead we split up the
`addNote` function into `addNote` and `addDraftNote`.
When `addNote` was called with `true`, we replace it by a call of `addDraftNote`
and add a check to see if draft notes can actually be added for the server version.

(cherry picked from commit 71293cb8c23c61bbf9b890c95dd9acfdcfd9e738)

IJ-CR-118475

GitOrigin-RevId: e97b9b275e27131150412df424da3a3fed40bcb1