IDEA-323662 Make the floating menu transparent when it's collapsed
The artifacts were caused by the menu being opaque and yet not
being painted. There's code in IdeJMenuBar.paint to avoid painting
a collapsed menu because it would paint a 1px line otherwise,
which is something that we don't need. However, for opaque
components it's not allowed to just skip painting, because
Swing assumes that if it's opaque, then it must paint over
the entire region it occupies.

Fix by simply making it non-opaque when the state changes to
COLLAPSED. This way Swing will happily paint the component
underneath it before calling the menu's paint.

GitOrigin-RevId: 5d9aa77cd3acc22ad2e57fb48b12f7acb249a49e