[file-history] reopen file history tabs when log is recreated
When project roots change or indexing gets enabled or disabled, VcsLogManager is recreated by VcsProjectLog. When this happens, log tabs are also recreated by the VcsLogTabsManager, but file history tabs are not. This can be especially confusing after user enables indexing from file history toolbar. In this commit, a FileHistoryTabsManager service is added to track file history tabs and to reopen them if necessary.

GitOrigin-RevId: e611591e6d6e1a80e41867c1d3636b699ca144cb