`blockingContext`: remove "rarely needed" clause from docs
`blockingContext` is expected to be used when calling any blocking code,
which the caller has no control over (e.g. 3rd party code).
In the beginning it was anticipated that the cancellation handle would be passed explicitly,
but it was proven inconvenient to drag the handle around everywhere.
The handle now resides in the thread local context (`Job`), which is filled
from the current coroutine context by this very `blockingContext` function.

This makes the clause not true anymore.

GitOrigin-RevId: aa139f3ef373f77ffa787c6248244043deff5e1a