IDEA-335860 Add speed search text placeholder
Since the field already supports displaying an empty text placeholder,
adding one is simple: just set the text. Reuse the same text as used
for the editor, as it fits here perfectly.

Use both the empty text and the actually entered text when
calculating the size of the field. This ensures that the field is large
enough to fit either and that its size doesn't jump when the
user starts to type: we use the maximum of both lengths,
so the field is always at least as large as needed to fit the empty text.

GitOrigin-RevId: 51011d9439c05445cdf1d87244ef218943e117fe