IDEA-335860 Add a blinking caret to speed search
It wasn't blinking before because the field is not focusable.
We want to keep it that way, but at the same time let the caret blink.
Fortunately, it's easy, as the caret is available through a public API
and has a method to explicitly set its visibility.

The existing Swing implementation resets its visibility in two cases:
when the field is disabled (never happens in this case) and when
the focus is lost, which doesn't happen here either because the field
is never focused to begin with, but just in case we override
the processFocusEvent method and again make the caret visible
there.

GitOrigin-RevId: 2a9383d9195b6a8e13fa74b37b0e663c697dfbd0