IDEA-323860 Work around for ActionMenu.fillMenu cancellation
The fillMenu() call in ActionMenu.setPopupMenuVisible() can produce
a cancellation exception (when expanding the action group).
Throwing exceptions during Swing event processing is likely
a bad idea in itself, but until we fix it, this simple workaround
will have to do. Remember the last passed boolean value
and check it when checking whether the currently selected menu
item is expanded in IdeJMenuBar.isActivated(), which is
used to determine whether the main menu should collapse
when hovering over the expanded items in full screen.

GitOrigin-RevId: ec633157ca9631885e25e2964cb8568bde573e9c