IDEA-336020 Fix editor scroll pane background color update
Two issues here.

First, it's often the case when both the old and the new value
of the background color are null. This corresponds to using
the default value from the scheme. However, since they are
both null, setting the property in EditorImpl.myState to null
doesn't fire the property change event, so the color isn't updated,
even though the new default may be different from the old default.

The other one is more tricky. This background color update
is triggered by the global LaF change when the theme is changed.
However, it doesn't guarantee that by that moment the editor color
scheme is already updated, so sometimes it ends up using the color
from the previous theme, so even forcing the callback for a null color
only leads to that the color now updates, but it's one cycle late:
for a dark theme it's light and for a light theme it's dark.

Fix by not relying on callbacks, and simply delegating
the scroll pane's background color to the editor's background color
using JBColor.lazy with a supplier. The downside is that the background
color for the scroll pane is always computed in a read action now,
but it happens exactly the same way for the editor component, so
it should not lead to any performance issues.

GitOrigin-RevId: 1794b6acf87ddb6321b964586309e7930e19a80e