[ui] Fix incorrect usages of JBUI#getInt
It returns the unscaled value from the UiDefaults, but sometimes the clients provide the scaled value as default - there is an inconsistency in a result. If a value is specified, we get it unscaled, otherwise we get scaled default.
Changed the usages to always return the unscaled value.
Also changed to return unscaled value from `List#rowHeight` and `Tree#rowHeight` to be consistent with other methods.

GitOrigin-RevId: ae91971b88e62f2314904ad8090d3679421435d7