IDEA-336046 Forward key type events to speed search when got it is shown
Now that we have a blinking caret in the speed search field,
it's definitely bad UX when it's impossible to type anything
while the got it tooltip is displayed.

There are two obvious ways to fix it: either don't focus the got it
tooltip anymore, but somehow still close it on enter, or keep it
focused but forward key events from it to the speed search field.
However, both approaches aren't easy to implement and have their
own drawbacks. In the first case, it's really hard for the speed search
field to close the got it tooltip, we'll have to provide additional API
for that, and then we'll have to use it in all speed search implementations.
In the second case, it's really hard to capture key events inside
the got it tooltip internals, as they're well encapsulated, as they should be.

So instead go the roundabout way and steal these events directly
from the event queue using a dispatcher that's disposed with the got it
tooltip. Only steal actual key typed events, so everything else is
processed as it should.

GitOrigin-RevId: 7f67eb2c65f93464be63fce23f06320ac69f0318